refactor(EmojiPicker): 将 EmojiPicker 移至 components/base 目录，对齐组件组织规范
