+++ src/renderer/components/base/EmojiPicker.tsx
